Refactor: Move AsyncProcess to worker_pool.py
Moved `AsyncProcess` out of `process_manager.py` into `worker_pool.py` to consolidate threading and process management into a single module. Updated imports in `plugin/modules/launcher/__init__.py` and `plugin/modules/tunnel/__init__.py`. Deleted `process_manager.py`.
